+++ source/extensions/omni.isaac.orbit_tasks/omni/isaac/orbit_tasks/manipulation/__init__.py
@@ -4,3 +4,5 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 """Manipulation environments for fixed-arm robots."""
+
+from .reach import *  # noqa